+++ haul-fe/src/views/Request/ChoiceLoadInfo/ChoiceLoadInfo.jsx
@@ -12,17 +12,9 @@ import styled from "styled-components";
 import NavigationBar from "../../../components/NavigationBar/NavigationBar.jsx";
 import BottomButton from "../../../components/Button/BottomButton.jsx";
 import FixedCenterBox from "../../../components/FixedBox/FixedCenterBox.jsx";
-import ToastMaker from "../../../components/Toast/ToastMaker.jsx";
-import {
-  UrlMap,
-  ErrorMessageMap,
-  TransportTypeArr
-} from "../../../data/GlobalVariable.js";
-import { isNumber, isPositiveNumber } from "../../../utils/helper.js";
-import { getIsMember } from "../../../utils/localStorage.js";
-import { memberReservationFun } from "../../../repository/reservationRepository.js";
-import { isTokenInvalid } from "../../../repository/userRepository.js";
+import { UrlMap } from "../../../data/GlobalVariable.js";
 import Loading from "../../Loading/Loading.jsx";
+import { SumbitStore, CheckSubmitDisabledFun } from "./index.jsx";
 
 const LoadInfoTypoBox = styled.div`
   width: 40px;
@@ -52,9 +44,9 @@ const SmallBtn = styled.button`
   text-align: center;
 `;
 
-const InputWrapper =styled.div`
+const InputWrapper = styled.div`
   width: 87%;
-`
+`;
 
 const ChoiceLoadInfo = () => {
   const navigate = useNavigate();
@@ -74,145 +66,25 @@ const ChoiceLoadInfo = () => {
     }
   } = useContext(reservationStore);
   const [submitDisabled, CheckSubmitDisabled] = useState(true);
-
-  useEffect(() => {
-    if (!dstAddress) {
-      //navigate(UrlMap.choiceDstPageUrl);
-    }
-    CheckSubmitDisabledFun();
-  }, []);
-
   const [inSpecialNotes, setInSpecialNotes] = useState(specialNotes);
-
   const inCargoWeight = useRef(cargoWeight.toString());
   const inCargoWidth = useRef(cargoWidth.toString());
   const inCargoLength = useRef(cargoLength.toString());
   const inCargoHeight = useRef(cargoHeight.toString());
 
-  //이 페이지에서 원하는 값이 다 있는지 체크
-  function CheckSubmitDisabledFun() {
-    inCargoWeight.current = inCargoWeight.current.trim();
-    inCargoWidth.current = inCargoWidth.current.trim();
-    inCargoLength.current = inCargoLength.current.trim();
-    inCargoHeight.current = inCargoHeight.current.trim();
-
-    const checkSubmitDisabled = !(
-      inCargoWeight.current &&
-      inCargoWidth.current &&
-      inCargoLength.current &&
-      inCargoHeight.current
-    );
-    if (submitDisabled !== checkSubmitDisabled) {
-      CheckSubmitDisabled(checkSubmitDisabled);
+  useEffect(() => {
+    if (!dstAddress) {
+      navigate(UrlMap.choiceDstPageUrl);
     }
-  }
-
-  async function SumbitStore() {
-    const cargoWeightNum = Number(inCargoWeight.current);
-    const cargoWidthNum = Number(inCargoWidth.current);
-    const cargoLengthNum = Number(inCargoLength.current);
-    const cargoHeightNum = Number(inCargoHeight.current);
-
-    if (
-      !isNumber(inCargoWeight.current) ||
-      !isNumber(inCargoWidth.current) ||
-      !isNumber(inCargoLength.current) ||
-      !isNumber(inCargoHeight.current)
-    ) {
-      ToastMaker({ type: "error", children: ErrorMessageMap.IsNotNumber });
-      return;
-    }
-    if (
-      !isPositiveNumber(inCargoWeight.current) ||
-      !isPositiveNumber(inCargoWidth.current) ||
-      !isPositiveNumber(inCargoLength.current) ||
-      !isPositiveNumber(inCargoHeight.current)
-    ) {
-      ToastMaker({
-        type: "error",
-        children: ErrorMessageMap.IsNotPositiveNumber
-      });
-      return;
-    }
-
-    function checkWeightToTypeFun(checkTransportType, weight) {
-      for (const transportInfo of TransportTypeArr) {
-        if (transportInfo.transportType === checkTransportType) {
-          if (weight > transportInfo.maxLoad * 1000) {
-            return transportInfo.maxLoad;
-          } else {
-            return 0;
-          }
-        }
-      }
-      return 0;
-    }
-
-    const checkWeightToType = checkWeightToTypeFun(
-      transportType,
-      cargoWeightNum
-    );
-    //transportType에 대한 분류
-    if (checkWeightToType > 0) {
-      ToastMaker({
-        type: "info",
-        children: `현재 선택하신 운송종류는 ${checkWeightToType}톤까지 실으실 수 있습니다.`
-      });
-      return;
-    }
-
-    //최대 너비, 높이, 길이에 대한 분류
-    if (
-      cargoWidthNum > 1000 ||
-      cargoLengthNum > 1000 ||
-      cargoHeightNum > 1000
-    ) {
-      ToastMaker({
-        type: "info",
-        children: `Haul은 1000cm미만의 크기를 지원합니다.`
-      });
-      return;
-    }
-
-    setRoadInfo({
-      cargoWeight: cargoWeightNum,
-      cargoWidth: cargoWidthNum,
-      cargoLength: cargoLengthNum,
-      cargoHeight: cargoHeightNum,
-      specialNotes: inSpecialNotes
+    CheckSubmitDisabledFun({
+      inCargoWeight: inCargoWeight.current,
+      inCargoWidth: inCargoWidth.current,
+      inCargoLength: inCargoLength.current,
+      inCargoHeight: inCargoHeight.current,
+      submitDisabled: submitDisabled,
+      CheckSubmitDisabled: CheckSubmitDisabled
     });
-
-    const isMember = getIsMember();
-    //비회원이라면 guestInfo 페이지로 이동
-    if (isMember === "false") {
-      navigate(UrlMap.guestInfoPageUrl);
-      return;
-    }
-    const reservationState = getReservationState();
-    // 회원이라면 바로 결과 페이지로 이동
-    setResultLoading(true);
-    const { success, data, code } = await memberReservationFun({
-      ...reservationState,
-      cargoWeight: cargoWeightNum,
-      cargoWidth: cargoWidthNum,
-      cargoLength: cargoLengthNum,
-      cargoHeight: cargoHeightNum,
-      specialNotes: inSpecialNotes
-    });
-    if (success) {
-      navigate(UrlMap.resultPageUrl, { state: { data: data.data } });
-    } else {
-      if (isTokenInvalid(code)) navigate(UrlMap.loginUrl);
-      if (code === 1104)
-        ToastMaker({
-          type: "error",
-          children: ErrorMessageMap.NoMatchingHaulCarError
-        });
-      else
-        ToastMaker({ type: "error", children: ErrorMessageMap.NetworkError });
-    }
-    setResultLoading(false);
-  }
+  }, []);
 
   if (resultLoading) return <Loading />;
 
@@ -239,7 +111,14 @@ const ChoiceLoadInfo = () => {
         defaultValue={cargoWeight}
         onChange={({ target: { value } }) => {
           inCargoWeight.current = value;
-          CheckSubmitDisabledFun();
+          CheckSubmitDisabledFun({
+            inCargoWeight: inCargoWeight.current,
+            inCargoWidth: inCargoWidth.current,
+            inCargoLength: inCargoLength.current,
+            inCargoHeight: inCargoHeight.current,
+            submitDisabled: submitDisabled,
+            CheckSubmitDisabled: CheckSubmitDisabled
+          });
         }}
         unit="kg"
         textAlign="right"
@@ -252,18 +131,24 @@ const ChoiceLoadInfo = () => {
       <Flex kind="flexBetweenAlignCenter">
         <LoadInfoTypoBox>너비</LoadInfoTypoBox>
         <InputWrapper>
-        <Input
-          size="small"
-          placeholder="짐의 너비를 알려주세요"
-          defaultValue={cargoWidth}
-          onChange={({ target: { value } }) => {
-            inCargoWidth.current = value;
-            CheckSubmitDisabledFun();
-          }}
-          unit="cm"
-          textAlign="right"
-
-        />
+          <Input
+            size="small"
+            placeholder="짐의 너비를 알려주세요"
+            defaultValue={cargoWidth}
+            onChange={({ target: { value } }) => {
+              inCargoWidth.current = value;
+              CheckSubmitDisabledFun({
+                inCargoWeight: inCargoWeight.current,
+                inCargoWidth: inCargoWidth.current,
+                inCargoLength: inCargoLength.current,
+                inCargoHeight: inCargoHeight.current,
+                submitDisabled: submitDisabled,
+                CheckSubmitDisabled: CheckSubmitDisabled
+              });
+            }}
+            unit="cm"
+            textAlign="right"
+          />
         </InputWrapper>
       </Flex>
 
@@ -271,17 +156,24 @@ const ChoiceLoadInfo = () => {
       <Flex kind="flexBetweenAlignCenter">
         <LoadInfoTypoBox>길이</LoadInfoTypoBox>
         <InputWrapper>
-        <Input
-          size="small"
-          placeholder="짐의 길이를 알려주세요"
-          defaultValue={cargoLength}
-          onChange={({ target: { value } }) => {
-            inCargoLength.current = value;
-            CheckSubmitDisabledFun();
-          }}
-          unit="cm"
-          textAlign="right"
-        />
+          <Input
+            size="small"
+            placeholder="짐의 길이를 알려주세요"
+            defaultValue={cargoLength}
+            onChange={({ target: { value } }) => {
+              inCargoLength.current = value;
+              CheckSubmitDisabledFun({
+                inCargoWeight: inCargoWeight.current,
+                inCargoWidth: inCargoWidth.current,
+                inCargoLength: inCargoLength.current,
+                inCargoHeight: inCargoHeight.current,
+                submitDisabled: submitDisabled,
+                CheckSubmitDisabled: CheckSubmitDisabled
+              });
+            }}
+            unit="cm"
+            textAlign="right"
+          />
         </InputWrapper>
       </Flex>
 
@@ -289,17 +181,24 @@ const ChoiceLoadInfo = () => {
       <Flex kind="flexBetweenAlignCenter">
         <LoadInfoTypoBox>높이</LoadInfoTypoBox>
         <InputWrapper>
-        <Input
-          size="small"
-          placeholder="짐의 높이를 알려주세요"
-          defaultValue={cargoHeight}
-          onChange={({ target: { value } }) => {
-            inCargoHeight.current = value;
-            CheckSubmitDisabledFun();
-          }}
-          unit="cm"
-          textAlign="right"
-        />
+          <Input
+            size="small"
+            placeholder="짐의 높이를 알려주세요"
+            defaultValue={cargoHeight}
+            onChange={({ target: { value } }) => {
+              inCargoHeight.current = value;
+              CheckSubmitDisabledFun({
+                inCargoWeight: inCargoWeight.current,
+                inCargoWidth: inCargoWidth.current,
+                inCargoLength: inCargoLength.current,
+                inCargoHeight: inCargoHeight.current,
+                submitDisabled: submitDisabled,
+                CheckSubmitDisabled: CheckSubmitDisabled
+              });
+            }}
+            unit="cm"
+            textAlign="right"
+          />
         </InputWrapper>
       </Flex>
 
@@ -331,7 +230,18 @@ const ChoiceLoadInfo = () => {
           role="main"
           disabled={submitDisabled}
           onClick={() => {
-            SumbitStore();
+            SumbitStore({
+              inCargoWeight: inCargoWeight,
+              inCargoWidth: inCargoWidth,
+              inCargoLength: inCargoLength,
+              inCargoHeight: inCargoHeight,
+              setRoadInfo: setRoadInfo,
+              navigate: navigate,
+              setResultLoading: setResultLoading,
+              transportType: transportType,
+              inSpecialNotes: inSpecialNotes,
+              getReservationState: getReservationState
+            });
           }}
         >
           선택 완료
